♻️ Tree 컴포넌트 리팩토링
recoil 관련 변수 tree에서 가져오도록 변경
